audit: add audit_health daemon tool; route gortex audit through it
The complexity-axis health computation moves to a new audit_health MCP
tool (ComputeAuditReport on the daemon's graph); gortex audit calls it,
decodes the report, and renders the badge / JSON / text and writes the
badge file locally. The computation tests move to internal/mcp with the
logic; the rendering tests stay. No daemon / untracked repo -> the
start+track error.
