+++ internal/mcp/server.go
@@ -996,6 +996,7 @@ func NewServer(engine *query.Engine, g graph.Store, idx *indexer.Indexer, watche
 	s.registerGraphCompletionTool()
 	s.registerWikiTools()
 	s.registerExportTools()
+	s.registerAuditTool()
 	s.registerWalkGraphTool()
 	s.registerContextClosureTool()
 	s.registerGraphQueryTool()
